Aggressive performance optimization
Subdivide event taps and switch between mutating and passive taps when possible. Passive (listen-only) taps are better for performance. The system don’t have to wait for the callback and can handle the event immediately. Now we use the passive tap for standard scroll wheel events, which increases the performance by about 80%.